fix shm in linux
